Add broker to server to fix TTL issues
If a command is queued longer than the munge TTL, then it will fail to
decode. This effectively defeats the advantage of queuing when
long-running commands are involved. Adding a broker creates a second
queue to resolve this. The front-end queue is used to receive and decode
messages. They are then placed in the backend queue where they are
processed and executed.

This broker also opens the door for multiple workers, potentially on
remote systems.